extract login component, use on Matches page only
