+++ src/App.vue
@@ -1,49 +1,10 @@
 <script setup lang="ts">
 import { RouterLink, RouterView } from 'vue-router';
-import { ref } from 'vue';
-import axios from 'axios';
-
-import {
-  setUserInStorage,
-  removeUserInStorage,
-  type ZugUser,
-} from '@/utils/auth';
-import { store } from '@/store';
-
-const usernameInput = ref('');
-
-const login = async () => {
-  const resp = await axios.post('/api/login', {
-    username: usernameInput.value,
-  });
-  if (resp.status === 200) {
-    const { data } = resp;
-    const { authToken, userID } = data as ZugUser;
-    setUserInStorage(data);
-    store.setZugToken(authToken);
-    store.setZugUsername(userID);
-  }
-};
-
-const logout = () => {
-  removeUserInStorage();
-  store.setZugToken(undefined);
-  store.setZugUsername(undefined);
-};
 </script>
 
 <template>
   <header>
     <div class="wrapper">
-      <div v-if="store.zugToken">
-        <p>hi {{ store.zugUsername }}!</p>
-        <button @click="logout">logout</button>
-      </div>
-      <div v-else>
-        <input class="login" v-model="usernameInput" />
-        <button @click="login">login</button>
-      </div>
-
       <nav>
         <RouterLink to="/">Home</RouterLink>
         <RouterLink to="/how-to-play">How To Play</RouterLink>
@@ -76,9 +37,5 @@ nav a {
 
 nav a:first-of-type {
   border: 0;
-}
-
-.login {
-  width: 8rem;
 }
 </style>
