+++ src/views/HomeView.vue
@@ -1,14 +1,11 @@
 <script setup lang="ts">
-import { RouterLink } from 'vue-router';
 import LobbyComponent from '@/components/LobbyComponent.vue';
-const { hostname } = window.location;
+import LoginComponent from '@/components/LoginComponent.vue';
 </script>
 
 <template>
   <main>
-    <RouterLink v-if="hostname === 'localhost'" to="/match/1"
-      >Go to Match #1</RouterLink
-    >
+    <LoginComponent />
     <LobbyComponent></LobbyComponent>
   </main>
 </template>
